Implement Device Code Form component and refactor Device Authorization page to utilize it. Update auth client base URL to support environment variables and enhance CORS configuration to include server URL.
