+++ apps/supercode-cli/client/lib/auth-client.ts
@@ -1,7 +1,7 @@
 import { createAuthClient } from "better-auth/react";
 import {deviceAuthorizationClient} from 'better-auth/client/plugins'
 export const authClient = createAuthClient({
-    baseURL: "http://localhost:3004",
+    baseURL: process.env.NEXT_PUBLIC_AUTH_URL || "http://localhost:3004",
     plugins:[
         deviceAuthorizationClient()
     ]
